backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-generativeai>=0.3.0
yt-dlp>=2023.12.30
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
